_content: move security to doc/security, with redirects
go.dev/security made sense as a URL when it was the security policy,
which might be considered not quite user-facing documentation.
But now it has grown into lots of definitely user-facing documentation,
so it should be in /doc with all the other user docs.

Move /security to /doc/security with redirects.

Change-Id: I4747bbd0d1c110a4c7f0ae2fe0098326cc8637b8
Reviewed-on: https://go-review.googlesource.com/c/website/+/548119
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Roland Shoemaker <roland@golang.org>
Auto-Submit: Russ Cox <rsc@golang.org>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>