Large outputs now show vertical scrollbar
The calc(100% - 52px) corrects for the 32px that the .output element sticks out below the bottom of the page, and 20px for for the 10px padding that all pre tags have on the top and bottom. If this isn't corrected for then the vertical scrollbar is rendered below the bottom of the page.