doc: unify terminology to Standard library
This aligns the existing release notes with the upcoming release notes for Go
1.24 and newer (https://go.dev/cl/556817).

The term Standard library is more common, both in spoken and written
communication about Go. Also, it is engrained in commands like
go test std, which is not going to change.

Change-Id: I5cc2bfc608effe40befa8f168b1f08b9882ce7dd
Reviewed-on: https://go-review.googlesource.com/c/website/+/598096
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>