_content/blog/range-functions: suggest go get over go mod edit
Using 'go mod edit -go=1.23' is equivalent to editing the go directive
in go.mod, and nothing more. This doesn't pay attention to the rest of
go.mod content and may leave the module in an untidy state.

Prior to Go 1.21, 'go mod tidy -go=1.23' was the suggested way to do
the update while also maintaining the tidy state of the module. That
form is still valid, but the simpler go get go@1.23 is now preferred¹.

The 'go get go@1.n' syntax is newer and not everyone has seen it yet,
so keep the 'go mod edit' mention with a hint as to how they differ.

¹ https://go.dev/doc/toolchain#get, second last paragraph

Change-Id: I5c8c5568e1512fec4de28e666a6b80be98290647
Reviewed-on: https://go-review.googlesource.com/c/website/+/607036
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Ian Lance Taylor <iant@google.com>