_content/doc/go1.23: drop link to rangefunc wiki, add blog post
The RangefuncExperiment wiki page has been updated to reflect the fact
that range-over-func is in 1.23, and the wiki page is no longer a good
destination for describing the change. Drop the link to it.
Add a link to the recent blog post.

Change-Id: I899f36f4172182b7089f4964e426530fd1865778
Reviewed-on: https://go-review.googlesource.com/c/website/+/608819
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Ian Lance Taylor <iant@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Ian Lance Taylor <iant@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>