internal/screentest: deprecate some unused features
This is the first of a series of CLs that aims to move screentest to a
more traditional testing model.

We can make these changes, which will ultimately be breaking,
because searches of the open source ecosystem and internal Google
code show that only this repo and x/pkgsite use screentest.
Also, this package is internal, this repo is at major version 0,
and the Go backward compatibility guarantee does not extend to command
line tools.

There are two types of changes I will be making.
First, the test files will contain only tests, not information about
what systems are being tested or where that information is stored.
This CL moves toward that goal by deprecating the "output" command
and requiring that certain other commands occur only inside tests.
Future CLs will deprecate more unused features, like multiple
"compare" commands in a single file. The ultimate goal is to
have the systems to be compared provided as input, along with
additional information like where to write output and what HTTP
headers to send.

The second, conceptually larger change is to move to a traditional
test model where one source is a reference or "golden," and the
other is the system under test. The current caching mechanism
can achieve this and more, but at the cost of complexity and confusion.
There are no active uses of caching as an actual caching mechanism,
only as a way to store and update goldens. This CL makes no progress
toward this goal, but I provide it here to lay out the full story.

To summarize the changes in this CL:

- Deprecate "output" in preference to a command-line argument.
- Deprecate use of many commands outside of tests, including the
  "pathname" command in lieu of a "test" declaration.
- Fix a bug (presumably; there are no uses) where the test filename
  is not part of the GCS destination URL.

Change-Id: Ibeffe62ba17d10265f7960a6ce0c380c4374c141
Reviewed-on: https://go-review.googlesource.com/c/website/+/626956
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>