cmd/screentest: replace caching with got/want model
Remove the '::caching' suffix from URLs. Instead of caching one
or both arguments to the filesystem or GCS bucket, a user can specify
a directory or bucket as one of the URL arguments. Most commonly,
the testURL will refer to a server and the wantURL will refer
to a set of "golden" files. The update flag will then overwrite
the goldens with the test results, or record them if there are
no goldens.

A useful side benefit is the ability to separate the location
of the goldens from the location where the output of failed tests is
written. For example, the former can be a directory and the latter
a GCS bucket.

This change also simplifies the code, since for the most part
locations, both test, golden and output, are treated uniformly.
There is still a distinction between servers and storage because
servers use a URL path to identify requires while storage uses a
path constructed from the test name, but reads and writes from
storage can be treated similarly.

Change-Id: I1136fa3ea8dbe3a3019594c260f6469cc7021891
Reviewed-on: https://go-review.googlesource.com/c/website/+/627118
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>