cmd/golangorg/testdatascreentest: fix learn test
Remove the Go version HTML from the screenshot so the test will
pass. The version mismatch is an artifact of the test setup.

Also, have screentest warn about using eval with quoted expressions.
I was doing this and couldn't understand why my fix to the learn
test wasn't working.

Lastly, describe how we could add a feature that shows the result
of evaluating each expression. My buggy version of this feature
helped me diagnose the above problem, but a proper implementation
is harder and not worth it now.

Change-Id: I411717cd2aafab97193082a392bcd792bd3c7534
Reviewed-on: https://go-review.googlesource.com/c/website/+/631715
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>