_content/doc/modules: clarify that tool's name isn't always basename
The name that can be passed to go tool to invoke a module-defined tool
isn't always the basename of the tool's import path. Change the wording
to clarify that.

This is the wording that we use in the go build documentation for where
we place its output by default. It's a bit awkward to have the phrase
here but I think it's okay

Change-Id: Idfab92c7504e4a01382345441a4aef9c35e4ba56
Reviewed-on: https://go-review.googlesource.com/c/website/+/634697
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Sam Thanawalla <samthanawalla@google.com>
Reviewed-by: Will Faught <will.faught@gmail.com>