_content/doc/go1.24: extend, fix, and cleanup crypto release notes
FIPS 140 commented out for now, will add a section in a follow-up CL.
It's ok for those to come after rc1, also because it's a bit unclear yet
what we can assert about the FIPS 140-3 status of the release. Note that
none of the changes affect existing programs that don't opt in with the
new switches.

For golang/go#68545

Change-Id: I9779a0976127e5fdc65b4b8ae3c6c2de4739390b
Reviewed-on: https://go-review.googlesource.com/c/website/+/635595
Auto-Submit: Filippo Valsorda <filippo@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Roland Shoemaker <roland@golang.org>
Reviewed-by: Daniel McCarney <daniel@binaryparadox.net>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>