_content/doc/gc-guide: set width of counter for diagram controls
Currently the diagram controls jitter when you move the slider, because
the width of the counter changes as you move the slider, and the element
is centered. Give the counter a fixed width to prevent this. The counter
never contains more than 10 characters (really 9, but let's give it some
wiggle room) so size it accordingly. This makes the element slightly
off-center, but it's hardly noticable.

Change-Id: I393d4fd06898aa034d0ce7fa573d2e484a164e1d
Reviewed-on: https://go-review.googlesource.com/c/website/+/637275
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>