_content/doc/go1.24: fix meaning of -u in 'go get -u tool'
It's a common misperception that -u is needed more often than it is,
perhaps in part due to muscle memory from the GOPATH days.

The current release notes suggest that 'go get -u tool' will upgrade all
of the current module's tools, but it does more than that -- it will
also upgrade all of the direct and indirect non-test dependencies of
those tools too.

The simplest fix is probably just to remove the -u from this section of
the release notes.

There are various other details that could be covered, like the -u or -t
flags, or that 'go get tool' is a shorthand for 'go get tool@upgrade',
but maybe this is sufficient.

Updates golang/go#48429
Updates golang/go#68545

Change-Id: I31eeb96dae996a7ffde24d9bae6788d6889e2044
Reviewed-on: https://go-review.googlesource.com/c/website/+/638296
Reviewed-by: David Chase <drchase@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Conrad Irwin <conrad.irwin@gmail.com>