_content/doc/go1.24: dedicated subsection for AddCleanup
AddCleanup is a new library routine. I think the fact that it's in the
runtime package is largely irrelevant to the user, so this CL moves
the documentation on AddCleanup out of the runtime section and into a
new subsection of "Standard library". It also tidies up the text.

Change-Id: I0609f48d3043bbd84bb0ef1e8c21f44baf8f1764
Reviewed-on: https://go-review.googlesource.com/c/website/+/638560
Reviewed-by: Ian Lance Taylor <iant@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>