_content/doc/go1.24: tweak os.Root section
Currently, this section links to four methods of os.Root in some
detail, without making it clear why it's linking to these methods or
that this is just a subset of the methods on os.Root. Clarify that the
methods of os.Root generally mirror the whole os file system API and
that we're linking to just a few supporting examples.

Change-Id: I8efb20a9b3946eb32d5695d9d955f8780e059c84
Reviewed-on: https://go-review.googlesource.com/c/website/+/638562
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Ian Lance Taylor <iant@google.com>
Reviewed-by: Damien Neil <dneil@google.com>