tour: deprecate and retract the nested module
It stopped being a nested module and its content was merged into
x/website in CL 323897. It was never intended to be a published
module, rather it happened to be one during earlier development
of x/website.

Because the golang.org/x/website/tour module has a longer path than
the golang.org/x/website module, it gets in the way when trying to
install or run the latest version of the tour command. So, deprecate
and retact all of its versions to fix that.

For golang/go#47236.

Change-Id: I88615fc25cd6531d4650210ca8063c53a68ac81a