_content/doc/go1.25: clean up some TODOs
A number of TODOs are for proposals implemented in previous
releases, just got updated with CLs in this cycle. Perhaps the
relnote tool should not pick up proposals with an old milestone
or closed since the beginning of the cycle.

Also delete the contents, which are not shown anyway.

For golang/go#71661.

Change-Id: Iaf7bef23dc93c4c406bcba3e2013072e36eaea7e
Reviewed-on: https://go-review.googlesource.com/c/website/+/678675
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>