_content/blog: add Go 1.25 release post
This post tries to be short. It lets readers know that Go 1.25 is out,
and as much as possible directs them to read the Go 1.25 Release Notes
to see what's new. That document is where we've invested most of the
effort of documenting and highlighting the Go 1.25 release changes,
so the blog post can mostly link to it.

It was tempting to try and include a subset of highlights, but it's
tricky to do that without either making it seem like it's a complete
list (when it's not), or making the post too long and harder to see
parts of it that that aren't already in scope of the release notes.

Balance this out by picking a handful of interesting features to
highlight in the blog post summary (where it's clear that it's not
the full list), and by enumerating experimental additions that would
benefit from early testing.

Change-Id: Iaca7c3a638a00a406e230306a0bff741c909a06e
Reviewed-on: https://go-review.googlesource.com/c/website/+/694835
Reviewed-by: David Chase <drchase@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Austin Clements <austin@google.com>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>