_content/tour: point to contribution guide for reporting issues
The feedback button makes it a bit too easy to file a new issue without
being prompted to first check if same issue has already been reported.

Try pointing to the Go contribution guide and its "check the issue
tracker" section, and see how that goes. The intent is that if someone
wishes to provide feedback, it's still possible to do that, and
hopefully the less immediate path to the issue tracker makes it more
likely that reporters prefer to find (and upvote and/or subscribe to)
an existing issue whenever it already exists.

Change-Id: I2343711530cf89982fb8725a791b19b5ee5e4a14
Reviewed-on: https://go-review.googlesource.com/c/website/+/708756
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>